requirements.txt
aiocontextvars==0.2.2
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextvars==2.4
cryptography==3.2.1
dataclasses==0.8
fastapi==0.48.0
h11==0.9.0
idna==2.8
immutables==0.14
joblib==0.14.1
loguru==0.4.1
numpy==1.18.1
passlib==1.7.4
pycparser==2.20
pydantic==1.7.2
PyJWT==1.7.1
python-multipart==0.0.5
requests==2.22.0
scikit-learn==0.22.1
scipy==1.1.0
six==1.15.0
starlette==0.12.9
urllib3==1.25.11
uvicorn==0.11.7
websockets==8.1
win32-setctime==1.0.3
wincertstore==0.2
